Created tests for detecting victory.
