+++ js/create.js
@@ -22,20 +22,8 @@ for(var i = 0; i < x; i++) {
       }
       
   }
-  console.log(board);
   return board;
 }
 
-function iswin(board){
-    let base = board[0][0]
-    for(var i = 0; i < x; i++) {
-        for(var j = 0; j < y; j++) {
-           if(board[i][j] != base){
-                return false;
-           }; 
-          }
-      }
-      return true;
-}
 
 module.exports = create;
